+++ modules/core/src/main/kotlin/io/github/lmliam/kotventure/core/selector/SelectorToken.kt
@@ -1,0 +1,10 @@
+package io.github.lmliam.kotventure.core.selector
+
+internal fun Char.isAllowedInUnquotedSelectorToken(): Boolean =
+    this in '0'..'9' ||
+        this in 'A'..'Z' ||
+        this in 'a'..'z' ||
+        this == '_' ||
+        this == '-' ||
+        this == '.' ||
+        this == '+'